requirements.txt
alembic==1.7.7
appdirs==1.4.3
apturl==0.5.2
argcomplete==2.0.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
django-cors-headers==3.12.0
dnspython==2.2.1
dominate==2.6.0
duplicity==0.8.12.0
email-validator==1.2.1
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Reuploaded==1.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==1.1.0
Jinja2==2.11.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
newsapi-python==0.2.6
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2-binary==2.9.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
PyDrive==1.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyperclip==1.7.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.12
sqlparse==0.4.2
systemd-python==234
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
visitor==0.1.3
wadllib==1.3.3
Werkzeug==1.0.1
WTForms==3.0.1
xkit==0.0.0
zipp==3.8.0
